[C8] Split dimension.go by dimension family (#2225)
model/drawing/dimension.go was 1106 LOC, 2.2x the 500-line ceiling. Split
the drawing-dimension geometry by family, leaving the value type + text
decoration + collection in place; each file is now under the ceiling:

- dimension.go (318): the DrawingDimension value type, text decoration
  (overrides/tolerance/inspection/dual-unit), the DrawingDimensions
  collection and its lookup/edit.
- dimension_recompute.go (165): the shared attachment/reference resolution —
  Recompute, the recompute dispatch, dimensionBasis, and the shared
  measure/anchor/vertex-snap/glyph-segment helpers.
- dimension_linear.go (160): linear + baseline/chain sets.
- dimension_ordinate.go (93): ordinate set + leader glyph.
- dimension_radial.go (248): radius/diameter/arc-length + the circle/arc
  edge pickers.
- dimension_angular.go (215): angular between two straight edges + arc glyph.

Behaviour-preserving: every function body is byte-identical, only moved;
imports narrowed per file. The seven Find*ByKey attachment sites are moved
verbatim into their family files (guarding them through the collision-checked
resolver is the separate phase-2 C6-c task). Drawing corpus + archguard pass
before and after with no golden updates. Net delta of recognizers /
tolerance constants / fallback sites / type assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2225.

(cherry picked from commit 6cef082e429518b5d06606dc5c1994431a65c5d0)
