+++ model/drawing/dimension.go
@@ -4,14 +4,10 @@ package drawing
 
 import (
 	"fmt"
-	"math"
 	"strconv"
 
 	"oblikovati.org/api/contract"
 	"oblikovati.org/api/types"
-	"oblikovati.org/kernel/geom"
-	"oblikovati.org/kernel/hlr"
-	"oblikovati.org/kernel/topo"
 	gmath "oblikovati.org/math"
 	"oblikovati.org/model/collview"
 )
@@ -20,6 +16,11 @@ import (
 // two points on a drawing view. It attaches to two model vertices (by reference key), so it is
 // associative — on recompute it re-resolves the vertices against the current model, re-projects
 // them through the view and re-measures, and the value updates when the model size changes.
+//
+// This file holds the core: the DrawingDimension value type, its text decoration (overrides,
+// tolerance, inspection, dual-unit), the DrawingDimensions collection and its lookup/edit. The
+// per-family geometry (linear/angular/radial/ordinate) and the shared attachment/recompute path
+// live in the sibling dimension_*.go files (M48 #2225).
 
 // DrawingDimension is one linear dimension on a sheet: its measured value (true model size), the
 // view it annotates, the two attached model vertices, and the drawing curves of its glyph
@@ -257,261 +258,6 @@ func formatDimValue(v float64, decimals int) string {
 	return strconv.FormatFloat(v, 'f', decimals, 64)
 }
 
-// AddLinear adds a linear dimension on the named base view between two pick points (sheet mm),
-// each snapped to the nearest projected model vertex so the value tracks the model. dimType
-// selects the measured component; offset stands the dimension line off the points (sheet mm).
-func (ds *DrawingDimensions) AddLinear(name, viewName string, dimType types.DrawingDimensionType, x1, y1, x2, y2, offset float64) (*DrawingDimension, error) {
-	view, body, basis, err := ds.dimensionBasis(viewName)
-	if err != nil {
-		return nil, err
-	}
-	keyA, okA := nearestVertexKey(body, view, basis, x1, y1)
-	keyB, okB := nearestVertexKey(body, view, basis, x2, y2)
-	if !okA || !okB {
-		return nil, fmt.Errorf(errViewNoVertices, viewName)
-	}
-	return ds.addLinearFromKeys(name, viewName, dimType, keyA, keyB, offset), nil
-}
-
-// addLinearFromKeys creates a linear dimension between two already-resolved vertex keys.
-func (ds *DrawingDimensions) addLinearFromKeys(name, viewName string, dimType types.DrawingDimensionType, keyA, keyB []byte, offset float64) *DrawingDimension {
-	d := &DrawingDimension{name: ds.uniqueName(name), dimType: dimType, viewName: viewName, keyA: keyA, keyB: keyB, offset: offset}
-	ds.recompute(d)
-	ds.items = append(ds.items, d)
-	return d
-}
-
-// AddBaselineSet adds a baseline set: a linear dimension from the first pick point to each of the
-// others, stacked at increasing offsets (the datum-referenced common case). Each point is snapped
-// to the nearest projected model vertex, so every dimension stays associative.
-func (ds *DrawingDimensions) AddBaselineSet(viewName string, dimType types.DrawingDimensionType, points [][2]float64) ([]*DrawingDimension, error) {
-	keys, err := ds.snapPoints(viewName, points)
-	if err != nil {
-		return nil, err
-	}
-	const baseGap = 14.0
-	out := make([]*DrawingDimension, 0, len(keys)-1)
-	for i := 1; i < len(keys); i++ {
-		out = append(out, ds.addLinearFromKeys("", viewName, dimType, keys[0], keys[i], -baseGap*float64(i)))
-	}
-	return out, nil
-}
-
-// AddChainSet adds a chain set: a linear dimension between each consecutive pair of pick points,
-// all on one line (running dimensions). Each point is snapped to the nearest model vertex.
-func (ds *DrawingDimensions) AddChainSet(viewName string, dimType types.DrawingDimensionType, points [][2]float64) ([]*DrawingDimension, error) {
-	keys, err := ds.snapPoints(viewName, points)
-	if err != nil {
-		return nil, err
-	}
-	const chainOffset = -14.0
-	out := make([]*DrawingDimension, 0, len(keys)-1)
-	for i := 0; i+1 < len(keys); i++ {
-		out = append(out, ds.addLinearFromKeys("", viewName, dimType, keys[i], keys[i+1], chainOffset))
-	}
-	return out, nil
-}
-
-// AddSetForViewCorners dimensions a base view's bounding-box corners as a baseline or chain set —
-// the single-action dimension set (a full multi-pick set is a follow-up). Aligned measurement
-// gives non-degenerate values (width / diagonal / height).
-func (ds *DrawingDimensions) AddSetForViewCorners(viewName string, dimType types.DrawingDimensionType, baseline bool) ([]*DrawingDimension, error) {
-	view, ok := ds.views.ByName(viewName)
-	if !ok {
-		return nil, fmt.Errorf("drawing: no view %q to dimension", viewName)
-	}
-	minX, minY, maxX, maxY, ok := view.BoundsMM()
-	if !ok {
-		return nil, fmt.Errorf("drawing: view %q has no geometry to dimension", viewName)
-	}
-	pts := [][2]float64{{minX, minY}, {maxX, minY}, {maxX, maxY}, {minX, maxY}}
-	if baseline {
-		return ds.AddBaselineSet(viewName, dimType, pts)
-	}
-	return ds.AddChainSet(viewName, dimType, pts)
-}
-
-// snapPoints resolves each pick point to a model-vertex key on the view; it needs at least two
-// points and a resolvable model.
-func (ds *DrawingDimensions) snapPoints(viewName string, points [][2]float64) ([][]byte, error) {
-	if len(points) < 2 {
-		return nil, fmt.Errorf("drawing: a dimension set needs at least two points, got %d", len(points))
-	}
-	view, body, basis, err := ds.dimensionBasis(viewName)
-	if err != nil {
-		return nil, err
-	}
-	keys := make([][]byte, len(points))
-	for i, p := range points {
-		k, ok := nearestVertexKey(body, view, basis, p[0], p[1])
-		if !ok {
-			return nil, fmt.Errorf(errViewNoVertices, viewName)
-		}
-		keys[i] = k
-	}
-	return keys, nil
-}
-
-// AddOrdinateSet adds one ordinate dimension per point, each measuring the point's offset from the
-// shared datum along one axis (axisHorizontal = view-X, else view-Y), drawn as a leader to its
-// value with no dimension line. The datum and every point snap to the nearest projected model
-// vertex, so every ordinate stays associative.
-func (ds *DrawingDimensions) AddOrdinateSet(viewName string, axisHorizontal bool, datum [2]float64, points [][2]float64) ([]*DrawingDimension, error) {
-	if len(points) == 0 {
-		return nil, fmt.Errorf("drawing: an ordinate set needs at least one point, got 0")
-	}
-	view, body, basis, err := ds.dimensionBasis(viewName)
-	if err != nil {
-		return nil, err
-	}
-	datumKey, ok := nearestVertexKey(body, view, basis, datum[0], datum[1])
-	if !ok {
-		return nil, fmt.Errorf(errViewNoVertices, viewName)
-	}
-	out := make([]*DrawingDimension, 0, len(points))
-	for _, p := range points {
-		key, ok := nearestVertexKey(body, view, basis, p[0], p[1])
-		if !ok {
-			return nil, fmt.Errorf(errViewNoVertices, viewName)
-		}
-		out = append(out, ds.addOrdinateFromKeys(viewName, axisHorizontal, datumKey, key))
-	}
-	return out, nil
-}
-
-// addOrdinateFromKeys creates one ordinate dimension between an already-resolved datum and point.
-func (ds *DrawingDimensions) addOrdinateFromKeys(viewName string, axisHorizontal bool, datumKey, pointKey []byte) *DrawingDimension {
-	d := &DrawingDimension{name: ds.uniqueName(""), dimType: types.OrdinateDimension, viewName: viewName, keyA: datumKey, keyB: pointKey, axisHorizontal: axisHorizontal}
-	ds.recompute(d)
-	ds.items = append(ds.items, d)
-	return d
-}
-
-// AddRadial adds a radius or diameter dimension on the circular model edge nearest the pick point
-// (sheet mm) projected on the named base view; the value re-measures when the model changes.
-func (ds *DrawingDimensions) AddRadial(name, viewName string, dimType types.DrawingDimensionType, pickX, pickY float64) (*DrawingDimension, error) {
-	view, body, basis, err := ds.dimensionBasis(viewName)
-	if err != nil {
-		return nil, err
-	}
-	key, ok := nearestCircularEdgeKey(body, view, basis, pickX, pickY)
-	if !ok {
-		return nil, fmt.Errorf("drawing: view %q has no circular edge to dimension", viewName)
-	}
-	d := &DrawingDimension{name: ds.uniqueName(name), dimType: dimType, viewName: viewName, edgeKey: key}
-	ds.recompute(d)
-	ds.items = append(ds.items, d)
-	return d, nil
-}
-
-// AddArcLength adds an arc-length dimension on the circular/arc model edge nearest the pick point
-// (sheet mm) projected on the named base view; the value re-measures when the model changes.
-func (ds *DrawingDimensions) AddArcLength(name, viewName string, pickX, pickY float64) (*DrawingDimension, error) {
-	view, body, basis, err := ds.dimensionBasis(viewName)
-	if err != nil {
-		return nil, err
-	}
-	key, ok := nearestArcEdgeKey(body, view, basis, pickX, pickY)
-	if !ok {
-		return nil, fmt.Errorf("drawing: view %q has no circular or arc edge to dimension", viewName)
-	}
-	d := &DrawingDimension{name: ds.uniqueName(name), dimType: types.ArcLengthDimension, viewName: viewName, edgeKey: key}
-	ds.recompute(d)
-	ds.items = append(ds.items, d)
-	return d, nil
-}
-
-// AddRadialForEachCircle adds a radius or diameter dimension for every distinct circular edge in
-// the named base view — auto hole callouts. Coincident projections (a through-hole's two rims) are
-// dimensioned once. Returns how many it added.
-func (ds *DrawingDimensions) AddRadialForEachCircle(viewName string, dimType types.DrawingDimensionType) (int, error) {
-	view, body, basis, err := ds.dimensionBasis(viewName)
-	if err != nil {
-		return 0, err
-	}
-	seen := map[string]bool{}
-	added := 0
-	for _, e := range body.Edges() {
-		circle, ok := e.Geometry().(geom.Circle)
-		if !ok {
-			continue
-		}
-		c := view.place(hlr.ProjectPoint(basis, circle.Center))
-		key := fmt.Sprintf("%.1f/%.1f/%.2f", float64(c.X), float64(c.Y), circle.Radius)
-		if seen[key] {
-			continue
-		}
-		seen[key] = true
-		ds.addRadialForEdge(viewName, dimType, e.ReferenceKey())
-		added++
-	}
-	if added == 0 {
-		return 0, fmt.Errorf("drawing: view %q has no circular edges to dimension", viewName)
-	}
-	return added, nil
-}
-
-// addRadialForEdge appends a radial dimension on a specific circular edge (already resolved).
-func (ds *DrawingDimensions) addRadialForEdge(viewName string, dimType types.DrawingDimensionType, edgeKey []byte) {
-	d := &DrawingDimension{name: ds.uniqueName(""), dimType: dimType, viewName: viewName, edgeKey: edgeKey}
-	ds.recompute(d)
-	ds.items = append(ds.items, d)
-}
-
-// AddAngular adds an angular dimension between the two straight model edges nearest the pick
-// points (sheet mm) on the named base view; the measured angle re-derives when the model changes.
-func (ds *DrawingDimensions) AddAngular(name, viewName string, x1, y1, x2, y2 float64) (*DrawingDimension, error) {
-	view, body, basis, err := ds.dimensionBasis(viewName)
-	if err != nil {
-		return nil, err
-	}
-	keyA, okA := nearestStraightEdgeKey(body, view, basis, x1, y1)
-	keyB, okB := nearestStraightEdgeKey(body, view, basis, x2, y2)
-	if !okA || !okB || string(keyA) == string(keyB) {
-		return nil, fmt.Errorf("drawing: view %q needs two distinct straight edges for an angular dimension", viewName)
-	}
-	d := &DrawingDimension{name: ds.uniqueName(name), dimType: types.AngularDimension, viewName: viewName, edgeKey: keyA, edgeKeyB: keyB}
-	ds.recompute(d)
-	ds.items = append(ds.items, d)
-	return d, nil
-}
-
-// AddAngularForFirstCorner dimensions the angle between the first two non-parallel straight edges
-// in the named base view — a single-action corner/bevel-angle callout.
-func (ds *DrawingDimensions) AddAngularForFirstCorner(viewName string) (*DrawingDimension, error) {
-	_, body, basis, err := ds.dimensionBasis(viewName)
-	if err != nil {
-		return nil, err
-	}
-	view, _ := ds.views.ByName(viewName)
-	type seg struct {
-		key []byte
-		dir [2]float64
-	}
-	var segs []seg
-	for _, e := range body.Edges() {
-		if s, ok := straightEdgeOnSheet(body, view, basis, e.ReferenceKey()); ok {
-			segs = append(segs, seg{e.ReferenceKey(), s.dir})
-		}
-	}
-	for i := 0; i < len(segs); i++ {
-		for j := i + 1; j < len(segs); j++ {
-			if a := angleBetweenDeg(segs[i].dir, segs[j].dir); a > 1 && a < 179 {
-				return ds.addAngularEdges(viewName, segs[i].key, segs[j].key), nil
-			}
-		}
-	}
-	return nil, fmt.Errorf("drawing: view %q has no two non-parallel straight edges to dimension", viewName)
-}
-
-// addAngularEdges appends an angular dimension between two already-resolved straight edges.
-func (ds *DrawingDimensions) addAngularEdges(viewName string, keyA, keyB []byte) *DrawingDimension {
-	d := &DrawingDimension{name: ds.uniqueName(""), dimType: types.AngularDimension, viewName: viewName, edgeKey: keyA, edgeKeyB: keyB}
-	ds.recompute(d)
-	ds.items = append(ds.items, d)
-	return d
-}
-
 // MoveText nudges the named dimension's value text by (dx, dy) sheet millimetres — the drag-the-
 // text gesture, so overlapping dimensions can be separated for readability.
 func (ds *DrawingDimensions) MoveText(name string, dx, dy float64) {
@@ -531,455 +277,6 @@ func (ds *DrawingDimensions) MoveLine(name string, dx, dy float64) {
 	}
 	d.offset += dx*d.nx + dy*d.ny
 	ds.recompute(d)
-}
-
-// Recompute re-measures every dimension against the current model — the associativity path.
-func (ds *DrawingDimensions) Recompute() {
-	for _, d := range ds.items {
-		ds.recompute(d)
-	}
-}
-
-// recompute re-binds the dimension's geometry, re-measures and rebuilds the glyph. With no
-// resolvable view, model or geometry it clears the dimension.
-func (ds *DrawingDimensions) recompute(d *DrawingDimension) {
-	d.curves = nil
-	view, body, basis, err := ds.dimensionBasis(d.viewName)
-	if err != nil {
-		return
-	}
-	if d.retrievedFrom != "" {
-		ds.recomputeRetrieved(d, view, basis)
-		return
-	}
-	if isRadial(d.dimType) {
-		ds.recomputeRadial(d, view, body, basis)
-		return
-	}
-	if d.dimType == types.AngularDimension {
-		ds.recomputeAngular(d, view, body, basis)
-		return
-	}
-	if d.dimType == types.OrdinateDimension {
-		ds.recomputeOrdinate(d, view, body, basis)
-		return
-	}
-	if d.dimType == types.ArcLengthDimension {
-		ds.recomputeArcLength(d, view, body, basis)
-		return
-	}
-	ds.recomputeLinear(d, view, body, basis)
-}
-
-// recomputeOrdinate re-binds the datum (keyA) and measured point (keyB), measures the point's
-// view-X or view-Y offset from the datum (the running coordinate), then builds the leader glyph.
-func (ds *DrawingDimensions) recomputeOrdinate(d *DrawingDimension, view *DrawingView, body *topo.Body, basis hlr.View) {
-	datum, okA := body.FindVertexByKey(d.keyA)
-	point, okB := body.FindVertexByKey(d.keyB)
-	if !okA || !okB {
-		return
-	}
-	pd := hlr.ProjectPoint(basis, datum.Point())
-	pp := hlr.ProjectPoint(basis, point.Point())
-	if d.axisHorizontal {
-		d.valueMM = math.Abs(float64(pp.X-pd.X)) * cmToMM
-	} else {
-		d.valueMM = math.Abs(float64(pp.Y-pd.Y)) * cmToMM
-	}
-	d.text = d.decorate(formatDimValue(d.valueMM, ds.decimals()))
-	curves, mx, my, nx, ny := ordinateDimensionCurves(view, view.place(pp), d.axisHorizontal)
-	d.curves = curves
-	d.setTextAnchor(mx, my, nx, ny, 1)
-}
-
-// recomputeAngular re-binds the two straight edges, projects them and measures the angle between
-// them (degrees), then builds the arc glyph.
-func (ds *DrawingDimensions) recomputeAngular(d *DrawingDimension, view *DrawingView, body *topo.Body, basis hlr.View) {
-	a, okA := straightEdgeOnSheet(body, view, basis, d.edgeKey)
-	b, okB := straightEdgeOnSheet(body, view, basis, d.edgeKeyB)
-	if !okA || !okB {
-		return
-	}
-	d.valueDeg = angleBetweenDeg(a.dir, b.dir)
-	d.text = d.decorate(formatDimValue(d.valueDeg, ds.decimals()) + "°")
-	curves, mx, my, nx, ny := angularDimensionCurves(a, b)
-	d.curves = curves
-	d.setTextAnchor(mx, my, nx, ny, 1)
-}
-
-// recomputeLinear re-binds the two vertices, re-projects them and rebuilds the linear glyph.
-func (ds *DrawingDimensions) recomputeLinear(d *DrawingDimension, view *DrawingView, body *topo.Body, basis hlr.View) {
-	va, okA := body.FindVertexByKey(d.keyA)
-	vb, okB := body.FindVertexByKey(d.keyB)
-	if !okA || !okB {
-		return
-	}
-	p1 := hlr.ProjectPoint(basis, va.Point())
-	p2 := hlr.ProjectPoint(basis, vb.Point())
-	ds.buildLinearGlyph(d, view, p1, p2, measureMM(d.dimType, p1, p2))
-}
-
-// buildLinearGlyph measures the (already projected) endpoints, decorates the value and rebuilds the
-// linear dimension glyph and text anchor — the shared tail of a picked linear dimension and a
-// retrieved model dimension (#1991).
-func (ds *DrawingDimensions) buildLinearGlyph(d *DrawingDimension, view *DrawingView, p1, p2 gmath.Point2, valueMM float64) {
-	d.valueMM = valueMM
-	d.text = d.decorate(formatDimValue(valueMM, ds.decimals()))
-	s1, s2 := view.place(p1), view.place(p2)
-	ax, ay := dimensionAxis(d.dimType, s1, s2)
-	var mx, my float64
-	d.curves, mx, my = linearDimensionCurves(s1, s2, ax, ay, d.offset)
-	// Lift the text off the dimension line, on the side away from the measured geometry.
-	d.setTextAnchor(mx, my, -ay, ax, sign(d.offset))
-}
-
-// setTextAnchor records the dimension-line midpoint (mx, my) and the dimension line's unit
-// perpendicular nx,ny (the offset axis — line-drag follows it). It lifts the default text anchor
-// textGapMM off the line along liftSign*perpendicular (away from the measured geometry).
-func (d *DrawingDimension) setTextAnchor(mx, my, nx, ny, liftSign float64) {
-	if l := math.Hypot(nx, ny); l > 1e-9 {
-		nx, ny = nx/l, ny/l
-	}
-	d.nx, d.ny = nx, ny
-	d.anchorX, d.anchorY = mx+nx*liftSign*textGapMM, my+ny*liftSign*textGapMM
-}
-
-// isRadial reports whether a dimension type measures a circular edge (radius or diameter).
-func isRadial(t types.DrawingDimensionType) bool {
-	return t == types.RadiusDimension || t == types.DiameterDimension
-}
-
-// recomputeRadial re-binds the circular edge, re-measures its radius/diameter (the true model
-// size) and rebuilds the leader glyph.
-func (ds *DrawingDimensions) recomputeRadial(d *DrawingDimension, view *DrawingView, body *topo.Body, basis hlr.View) {
-	edge, ok := body.FindEdgeByKey(d.edgeKey)
-	if !ok {
-		return
-	}
-	circle, ok := edge.Geometry().(geom.Circle)
-	if !ok {
-		return
-	}
-	radiusMM := circle.Radius * cmToMM
-	if d.dimType == types.DiameterDimension {
-		d.valueMM = 2 * radiusMM
-		// Ø (U+00D8) is the diameter prefix: it is in the head's Latin-1 font, unlike the
-		// typographic ⌀ (U+2300), which renders as a missing-glyph box.
-		d.text = d.decorate("Ø" + formatDimValue(d.valueMM, ds.decimals()))
-	} else {
-		d.valueMM = radiusMM
-		d.text = d.decorate("R" + formatDimValue(d.valueMM, ds.decimals()))
-	}
-	center := view.place(hlr.ProjectPoint(basis, circle.Center))
-	arc := view.place(hlr.ProjectPoint(basis, circle.PointAt(0)))
-	opp := view.place(hlr.ProjectPoint(basis, circle.PointAt(0.5)))
-	curves, mx, my, nx, ny := radialDimensionCurves(center, arc, opp, d.dimType == types.DiameterDimension)
-	d.curves = curves
-	d.setTextAnchor(mx, my, nx, ny, 1)
-}
-
-// arcEdgeOf classifies an edge's geometry as a circular/arc curve and exposes what an arc-length
-// dimension needs: the centre, the swept length (cm), whether it is a full circle, and a sampler
-// that walks the curve as t runs 0→1. It reports ok=false for any non-circular edge.
-func arcEdgeOf(curve geom.Curve3) (center gmath.Point3, lengthCM float64, full bool, sample func(float64) gmath.Point3, ok bool) {
-	switch c := curve.(type) {
-	case geom.Circle:
-		return c.Center, c.Circumference(), true, c.PointAt, true
-	case geom.Arc3d:
-		return c.Center, c.Length(), false, c.PointAt, true
-	default:
-		return gmath.Point3{}, 0, false, nil, false
-	}
-}
-
-// recomputeArcLength re-binds the circular/arc edge, re-measures its swept length (the true model
-// size) and rebuilds the glyph — a dimension line following the projected arc.
-func (ds *DrawingDimensions) recomputeArcLength(d *DrawingDimension, view *DrawingView, body *topo.Body, basis hlr.View) {
-	edge, ok := body.FindEdgeByKey(d.edgeKey)
-	if !ok {
-		return
-	}
-	center3, lengthCM, full, sample, ok := arcEdgeOf(edge.Geometry())
-	if !ok {
-		return
-	}
-	d.valueMM = lengthCM * cmToMM
-	d.text = d.decorate(formatDimValue(d.valueMM, ds.decimals()))
-	const samples = 32
-	pts := make([]gmath.Point2, 0, samples+1)
-	for i := 0; i <= samples; i++ {
-		pts = append(pts, view.place(hlr.ProjectPoint(basis, sample(float64(i)/samples))))
-	}
-	centerS := view.place(hlr.ProjectPoint(basis, center3))
-	curves, mx, my, nx, ny := arcLengthDimensionCurves(centerS, pts, full)
-	d.curves = curves
-	d.setTextAnchor(mx, my, nx, ny, 1)
-}
-
-// arcLengthDimensionCurves builds an arc-length dimension's glyph (sheet mm): the projected arc
-// offset radially outward from the centre into a dimension line that follows it, with an arrowhead
-// at each end (a full circle has none). It returns the text anchor at the arc's middle, lifted
-// radially outward.
-func arcLengthDimensionCurves(center gmath.Point2, pts []gmath.Point2, full bool) (curves []DrawingCurve, anchorX, anchorY, nx, ny float64) {
-	const gap = 6.0
-	cx, cy := float64(center.X), float64(center.Y)
-	off := make([][2]float64, len(pts))
-	for i, p := range pts {
-		px, py := float64(p.X), float64(p.Y)
-		dx, dy := px-cx, py-cy
-		if l := math.Hypot(dx, dy); l > 1e-9 {
-			off[i] = [2]float64{px + dx/l*gap, py + dy/l*gap}
-		} else {
-			off[i] = [2]float64{px, py}
-		}
-	}
-	for i := 1; i < len(off); i++ {
-		curves = append(curves, dimSegment(off[i-1][0], off[i-1][1], off[i][0], off[i][1]))
-	}
-	if n := len(off); !full && n >= 2 {
-		curves = append(curves, arrowheadCurves(off[0][0], off[0][1], off[1][0], off[1][1])...)
-		curves = append(curves, arrowheadCurves(off[n-1][0], off[n-1][1], off[n-2][0], off[n-2][1])...)
-	}
-	mid := off[len(off)/2]
-	return curves, mid[0], mid[1], mid[0] - cx, mid[1] - cy
-}
-
-// nearestArcEdgeKey returns the reference key of the circular/arc model edge whose centre projects
-// nearest the pick (sheet mm), and false when the body has no such edges.
-func nearestArcEdgeKey(body *topo.Body, v *DrawingView, basis hlr.View, x, y float64) ([]byte, bool) {
-	var bestKey []byte
-	bestD := -1.0
-	for _, e := range body.Edges() {
-		center, _, _, _, ok := arcEdgeOf(e.Geometry())
-		if !ok {
-			continue
-		}
-		s := v.place(hlr.ProjectPoint(basis, center))
-		dx, dy := float64(s.X)-x, float64(s.Y)-y
-		if d := dx*dx + dy*dy; bestD < 0 || d < bestD {
-			bestD, bestKey = d, e.ReferenceKey()
-		}
-	}
-	return bestKey, bestD >= 0
-}
-
-// nearestCircularEdgeKey returns the reference key of the circular model edge whose centre projects
-// nearest the pick (sheet mm), and false when the body has no circular edges.
-func nearestCircularEdgeKey(body *topo.Body, v *DrawingView, basis hlr.View, x, y float64) ([]byte, bool) {
-	var bestKey []byte
-	bestD := -1.0
-	for _, e := range body.Edges() {
-		circle, ok := e.Geometry().(geom.Circle)
-		if !ok {
-			continue
-		}
-		s := v.place(hlr.ProjectPoint(basis, circle.Center))
-		dx, dy := float64(s.X)-x, float64(s.Y)-y
-		if d := dx*dx + dy*dy; bestD < 0 || d < bestD {
-			bestD, bestKey = d, e.ReferenceKey()
-		}
-	}
-	return bestKey, bestD >= 0
-}
-
-// radialDimensionCurves builds a radial dimension's glyph (sheet mm): a radius leader from the
-// centre to the arc with an arrowhead, or a diameter line across the circle with arrowheads at
-// both ends. It returns the text anchor (the midpoint of the leader / the centre).
-func radialDimensionCurves(center, arc, opp gmath.Point2, diameter bool) (curves []DrawingCurve, anchorX, anchorY, nx, ny float64) {
-	cx, cy := float64(center.X), float64(center.Y)
-	ax, ay := float64(arc.X), float64(arc.Y)
-	// Lift the text perpendicular to the leader/diameter line so it clears it.
-	pnx, pny := -(ay - cy), ax-cx
-	if diameter {
-		ox, oy := float64(opp.X), float64(opp.Y)
-		out := []DrawingCurve{dimSegment(ax, ay, ox, oy)}
-		out = append(out, arrowheadCurves(ax, ay, cx, cy)...)
-		out = append(out, arrowheadCurves(ox, oy, cx, cy)...)
-		return out, cx, cy, pnx, pny
-	}
-	out := []DrawingCurve{dimSegment(cx, cy, ax, ay)}
-	out = append(out, arrowheadCurves(ax, ay, cx, cy)...)
-	return out, (cx + ax) / 2, (cy + ay) / 2, pnx, pny
-}
-
-// sheetSegment is a straight model edge projected onto the sheet: its endpoints, midpoint and unit
-// direction (sheet mm).
-type sheetSegment struct {
-	ax, ay, bx, by float64
-	mx, my         float64
-	dir            [2]float64
-}
-
-// straightEdgeOnSheet resolves a straight edge by key and projects it onto the view's plane.
-func straightEdgeOnSheet(body *topo.Body, v *DrawingView, basis hlr.View, key []byte) (sheetSegment, bool) {
-	edge, ok := body.FindEdgeByKey(key)
-	if !ok {
-		return sheetSegment{}, false
-	}
-	line, ok := edge.Geometry().(geom.LineSegment)
-	if !ok {
-		return sheetSegment{}, false
-	}
-	a := v.place(hlr.ProjectPoint(basis, line.StartPoint))
-	b := v.place(hlr.ProjectPoint(basis, line.EndPoint))
-	ax, ay, bx, by := float64(a.X), float64(a.Y), float64(b.X), float64(b.Y)
-	dx, dy := bx-ax, by-ay
-	l := math.Hypot(dx, dy)
-	if l < 1e-9 {
-		return sheetSegment{}, false
-	}
-	return sheetSegment{ax, ay, bx, by, (ax + bx) / 2, (ay + by) / 2, [2]float64{dx / l, dy / l}}, true
-}
-
-// nearestStraightEdgeKey returns the reference key of the straight model edge whose midpoint
-// projects nearest the pick (sheet mm), and false when the body has no straight edges.
-func nearestStraightEdgeKey(body *topo.Body, v *DrawingView, basis hlr.View, x, y float64) ([]byte, bool) {
-	var bestKey []byte
-	bestD := -1.0
-	for _, e := range body.Edges() {
-		line, ok := e.Geometry().(geom.LineSegment)
-		if !ok {
-			continue
-		}
-		mid := line.StartPoint.Midpoint(line.EndPoint)
-		s := v.place(hlr.ProjectPoint(basis, mid))
-		dx, dy := float64(s.X)-x, float64(s.Y)-y
-		if d := dx*dx + dy*dy; bestD < 0 || d < bestD {
-			bestD, bestKey = d, e.ReferenceKey()
-		}
-	}
-	return bestKey, bestD >= 0
-}
-
-// angleBetweenDeg is the angle between two unit directions, in degrees ∈ [0, 180].
-func angleBetweenDeg(d1, d2 [2]float64) float64 {
-	dot := math.Max(-1, math.Min(1, d1[0]*d2[0]+d1[1]*d2[1]))
-	return math.Acos(dot) * 180 / math.Pi
-}
-
-// angularDimensionCurves builds an angular dimension's arc glyph (sheet mm) spanning between the
-// two edges, centred on their intersection, with arrowheads at the arc ends. It returns the text
-// anchor on the arc's bisector.
-func angularDimensionCurves(a, b sheetSegment) (curves []DrawingCurve, anchorX, anchorY, nx, ny float64) {
-	cx, cy, ok := lineIntersection(a, b)
-	if !ok {
-		cx, cy = (a.mx+b.mx)/2, (a.my+b.my)/2
-	}
-	const radius = 22.0
-	a0 := math.Atan2(a.my-cy, a.mx-cx)
-	sweep := normalizeAngle(math.Atan2(b.my-cy, b.mx-cx) - a0)
-	out := arcPolyline(cx, cy, radius, a0, a0+sweep)
-	out = append(out, arcArrowheads(cx, cy, radius, a0, a0+sweep)...)
-	mid := a0 + sweep/2
-	// The text sits on the bisector, lifted radially outward (away from the corner).
-	return out, cx + radius*math.Cos(mid), cy + radius*math.Sin(mid), math.Cos(mid), math.Sin(mid)
-}
-
-// lineIntersection returns the intersection of the two segments' infinite lines, ok=false when
-// they are parallel.
-func lineIntersection(a, b sheetSegment) (x, y float64, ok bool) {
-	denom := a.dir[0]*b.dir[1] - a.dir[1]*b.dir[0]
-	if math.Abs(denom) < 1e-9 {
-		return 0, 0, false
-	}
-	dx, dy := b.mx-a.mx, b.my-a.my
-	t := (dx*b.dir[1] - dy*b.dir[0]) / denom
-	return a.mx + t*a.dir[0], a.my + t*a.dir[1], true
-}
-
-// arcPolyline samples a circular arc (centre, radius, from angle a0 to a1) into drawing curves.
-func arcPolyline(cx, cy, r, a0, a1 float64) []DrawingCurve {
-	const n = 24
-	out := make([]DrawingCurve, 0, n)
-	px, py := cx+r*math.Cos(a0), cy+r*math.Sin(a0)
-	for i := 1; i <= n; i++ {
-		t := a0 + (a1-a0)*float64(i)/n
-		x, y := cx+r*math.Cos(t), cy+r*math.Sin(t)
-		out = append(out, dimSegment(px, py, x, y))
-		px, py = x, y
-	}
-	return out
-}
-
-// arcArrowheads builds an arrowhead at each end of the arc, pointing tangentially into the sweep.
-func arcArrowheads(cx, cy, r, a0, a1 float64) []DrawingCurve {
-	const eps = 0.12 // radians inward, to set the barb direction
-	tip0x, tip0y := cx+r*math.Cos(a0), cy+r*math.Sin(a0)
-	from0x, from0y := cx+r*math.Cos(a0+sign(a1-a0)*eps), cy+r*math.Sin(a0+sign(a1-a0)*eps)
-	tip1x, tip1y := cx+r*math.Cos(a1), cy+r*math.Sin(a1)
-	from1x, from1y := cx+r*math.Cos(a1-sign(a1-a0)*eps), cy+r*math.Sin(a1-sign(a1-a0)*eps)
-	out := arrowheadCurves(tip0x, tip0y, from0x, from0y)
-	return append(out, arrowheadCurves(tip1x, tip1y, from1x, from1y)...)
-}
-
-// normalizeAngle wraps an angle into (-π, π], so an arc sweeps the short way between two edges.
-func normalizeAngle(d float64) float64 {
-	for d > math.Pi {
-		d -= 2 * math.Pi
-	}
-	for d <= -math.Pi {
-		d += 2 * math.Pi
-	}
-	return d
-}
-
-func sign(v float64) float64 {
-	if v < 0 {
-		return -1
-	}
-	return 1
-}
-
-// dimensionBasis resolves a dimension's view (which must be a base view in this increment), the
-// referenced model body and the view's projection frame.
-func (ds *DrawingDimensions) dimensionBasis(viewName string) (*DrawingView, *topo.Body, hlr.View, error) {
-	view, ok := ds.views.ByName(viewName)
-	if !ok {
-		return nil, nil, hlr.View{}, fmt.Errorf("drawing: no view %q to dimension", viewName)
-	}
-	if view.viewType != types.DrawingViewBase {
-		return nil, nil, hlr.View{}, fmt.Errorf("drawing: %q is not a base view; dimension a base view", viewName)
-	}
-	if ds.body == nil {
-		return nil, nil, hlr.View{}, fmt.Errorf("drawing: no referenced model to dimension")
-	}
-	body, ok := ds.body()
-	if !ok {
-		return nil, nil, hlr.View{}, fmt.Errorf("drawing: no referenced model to dimension")
-	}
-	return view, body, baseBasis(view.orientation, bodyCenter(body)), nil
-}
-
-// measureMM is the dimension's value in millimetres from the two view-plane points (centimetres):
-// the true distance for an aligned dimension, or the view-X/Y component for horizontal/vertical.
-func measureMM(t types.DrawingDimensionType, p1, p2 gmath.Point2) float64 {
-	du, dv := float64(p2.X-p1.X), float64(p2.Y-p1.Y)
-	switch t {
-	case types.HorizontalDimension:
-		return math.Abs(du) * cmToMM
-	case types.VerticalDimension:
-		return math.Abs(dv) * cmToMM
-	default:
-		return math.Hypot(du, dv) * cmToMM
-	}
-}
-
-// dimensionAxis is the sheet-space direction the dimension is measured along: the line between the
-// points for an aligned dimension, or the sheet X/Y axis for horizontal/vertical.
-func dimensionAxis(t types.DrawingDimensionType, s1, s2 gmath.Point2) (ax, ay float64) {
-	switch t {
-	case types.HorizontalDimension:
-		return 1, 0
-	case types.VerticalDimension:
-		return 0, 1
-	default:
-		dx, dy := float64(s2.X-s1.X), float64(s2.Y-s1.Y)
-		l := math.Hypot(dx, dy)
-		if l < 1e-9 {
-			return 1, 0
-		}
-		return dx / l, dy / l
-	}
 }
 
 // Count, Item, ByName and Remove read/edit the collection.
@@ -1017,90 +314,5 @@ func (ds *DrawingDimensions) uniqueName(requested string) string {
 		if _, exists := ds.ByName(name); !exists {
 			return name
 		}
-	}
-}
-
-// nearestVertexKey returns the reference key of the model vertex whose projection on the view sits
-// closest to the pick (sheet mm), and false when the body has no vertices.
-func nearestVertexKey(body *topo.Body, v *DrawingView, basis hlr.View, x, y float64) ([]byte, bool) {
-	verts := body.Vertices()
-	best := -1
-	var bestD float64
-	for i, vt := range verts {
-		s := v.place(hlr.ProjectPoint(basis, vt.Point()))
-		dx, dy := float64(s.X)-x, float64(s.Y)-y
-		d := dx*dx + dy*dy
-		if best < 0 || d < bestD {
-			best, bestD = i, d
-		}
-	}
-	if best < 0 {
-		return nil, false
-	}
-	return verts[best].ReferenceKey(), true
-}
-
-// linearDimensionCurves builds a linear dimension's glyph (sheet mm): an extension line from each
-// measured point out to the dimension line, the dimension line itself, and an arrowhead at each
-// end. The dimension line is parallel to (ax, ay) and offset along its perpendicular by offset.
-// It also returns the dimension line's midpoint (the text anchor).
-func linearDimensionCurves(s1, s2 gmath.Point2, ax, ay, offset float64) (curves []DrawingCurve, anchorX, anchorY float64) {
-	nx, ny := -ay, ax // perpendicular to the measurement axis
-	s1x, s1y := float64(s1.X), float64(s1.Y)
-	s2x, s2y := float64(s2.X), float64(s2.Y)
-	lvl := s1x*nx + s1y*ny + offset // n-coordinate of the dimension line
-	e1x, e1y := s1x+offset*nx, s1y+offset*ny
-	d2 := lvl - (s2x*nx + s2y*ny)
-	e2x, e2y := s2x+d2*nx, s2y+d2*ny
-	out := []DrawingCurve{
-		dimSegment(s1x, s1y, e1x, e1y),
-		dimSegment(s2x, s2y, e2x, e2y),
-		dimSegment(e1x, e1y, e2x, e2y),
-	}
-	out = append(out, arrowheadCurves(e1x, e1y, e2x, e2y)...)
-	out = append(out, arrowheadCurves(e2x, e2y, e1x, e1y)...)
-	return out, (e1x + e2x) / 2, (e1y + e2y) / 2
-}
-
-// ordinateDimensionCurves builds an ordinate leader (sheet mm): a witness line from the measured
-// point out to a common spine — below the view for a horizontal (view-X) ordinate, to its right
-// for a vertical (view-Y) ordinate — so every ordinate text in a set lines up. It returns the text
-// anchor at the spine end and the outward lift direction. There is no dimension line or arrowhead.
-func ordinateDimensionCurves(view *DrawingView, point gmath.Point2, axisHorizontal bool) (curves []DrawingCurve, anchorX, anchorY, nx, ny float64) {
-	px, py := float64(point.X), float64(point.Y)
-	const spineGap = 14.0
-	_, minY, maxX, _, ok := view.BoundsMM()
-	if !ok {
-		minY, maxX = py, px
-	}
-	if axisHorizontal {
-		spineY := minY - spineGap
-		return []DrawingCurve{dimSegment(px, py, px, spineY)}, px, spineY, 0, -1
-	}
-	spineX := maxX + spineGap
-	return []DrawingCurve{dimSegment(px, py, spineX, py)}, spineX, py, 1, 0
-}
-
-// arrowheadCurves builds the two short barbs of an arrowhead whose tip is at (tipX, tipY) and
-// whose shaft runs toward (fromX, fromY).
-func arrowheadCurves(tipX, tipY, fromX, fromY float64) []DrawingCurve {
-	const length, angle = 3.0, 0.35 // mm, radians (~20°)
-	ux, uy := fromX-tipX, fromY-tipY
-	l := math.Hypot(ux, uy)
-	if l < 1e-9 {
-		return nil
-	}
-	ux, uy = ux/l, uy/l
-	cos, sin := math.Cos(angle), math.Sin(angle)
-	b1x, b1y := tipX+length*(ux*cos-uy*sin), tipY+length*(ux*sin+uy*cos)
-	b2x, b2y := tipX+length*(ux*cos+uy*sin), tipY+length*(-ux*sin+uy*cos)
-	return []DrawingCurve{dimSegment(tipX, tipY, b1x, b1y), dimSegment(tipX, tipY, b2x, b2y)}
-}
-
-// dimSegment builds a visible drawing curve between two sheet-mm points.
-func dimSegment(ax, ay, bx, by float64) DrawingCurve {
-	return DrawingCurve{
-		A: gmath.P2(gmath.Scalar(ax), gmath.Scalar(ay)), B: gmath.P2(gmath.Scalar(bx), gmath.Scalar(by)),
-		Visible: true, kind: types.DrawingEdgeCurve,
 	}
 }
